+++ engine/include/genebits/engine/scheduler/stage.h
@@ -1,8 +1,8 @@
-#ifndef GENEBITS_ENGINE_ECS_STAGE_H
-#define GENEBITS_ENGINE_ECS_STAGE_H
+#ifndef GENEBITS_ENGINE_SCHEDULER_STAGE_H
+#define GENEBITS_ENGINE_SCHEDULER_STAGE_H
 
 #include "genebits/engine/containers/vector.h"
-#include "genebits/engine/ecs/system.h"
+#include "genebits/engine/system/system.h"
 #include "genebits/engine/utilities/ref.h"
 
 namespace genebits::engine
