+++ engine/include/genebits/engine/system/context.h
@@ -1,5 +1,5 @@
-#ifndef GENEBITS_ENGINE_ECS_CONTEXT_H
-#define GENEBITS_ENGINE_ECS_CONTEXT_H
+#ifndef GENEBITS_ENGINE_SYSTEM_CONTEXT_H
+#define GENEBITS_ENGINE_SYSTEM_CONTEXT_H
 
 #include <bit>
