+++ engine/test/ecs/scheduler_algorithm_tests.cpp
@@ -1,4 +1,4 @@
-#include "genebits/engine/ecs/scheduler.h"
+#include "genebits/engine/scheduler/scheduler.h"
 
 #include <gtest/gtest.h>
