+++ sandbox/systems/main.cpp
@@ -4,7 +4,7 @@
 #include "genebits/engine/async/sync_wait.h"
 #include "genebits/engine/async/task.h"
 #include "genebits/engine/debug/logging.h"
-#include "genebits/engine/ecs/system.h"
+#include "genebits/engine/system/system.h"
 
 using namespace genebits::engine;
